chore: refactor responsive style
