+++ src/layouts/RootLayout/index.tsx
@@ -1,44 +1,43 @@
-import React, { ReactNode, useEffect } from "react"
-import { ThemeProvider } from "./ThemeProvider"
-import useScheme from "src/hooks/useScheme"
-import Header from "./Header"
 import styled from "@emotion/styled"
-import Scripts from "src/layouts/RootLayout/Scripts"
-import useGtagEffect from "./useGtagEffect"
 import Prism from "prismjs/prism"
-import 'prismjs/components/prism-markup-templating.js'
-import 'prismjs/components/prism-markup.js'
-import 'prismjs/components/prism-bash.js'
-import 'prismjs/components/prism-c.js'
-import 'prismjs/components/prism-cpp.js'
-import 'prismjs/components/prism-csharp.js'
-import 'prismjs/components/prism-docker.js'
-import 'prismjs/components/prism-java.js'
-import 'prismjs/components/prism-js-templates.js'
-import 'prismjs/components/prism-coffeescript.js'
-import 'prismjs/components/prism-diff.js'
-import 'prismjs/components/prism-git.js'
-import 'prismjs/components/prism-go.js'
-import 'prismjs/components/prism-kotlin.js'
-import 'prismjs/components/prism-graphql.js'
-import 'prismjs/components/prism-handlebars.js'
-import 'prismjs/components/prism-less.js'
-import 'prismjs/components/prism-makefile.js'
-import 'prismjs/components/prism-markdown.js'
-import 'prismjs/components/prism-objectivec.js'
-import 'prismjs/components/prism-ocaml.js'
-import 'prismjs/components/prism-python.js'
-import 'prismjs/components/prism-reason.js'
-import 'prismjs/components/prism-rust.js'
-import 'prismjs/components/prism-sass.js'
-import 'prismjs/components/prism-scss.js'
-import 'prismjs/components/prism-solidity.js'
-import 'prismjs/components/prism-sql.js'
-import 'prismjs/components/prism-stylus.js'
-import 'prismjs/components/prism-swift.js'
-import 'prismjs/components/prism-wasm.js'
-import 'prismjs/components/prism-yaml.js'
+import "prismjs/components/prism-bash.js"
+import "prismjs/components/prism-c.js"
+import "prismjs/components/prism-coffeescript.js"
+import "prismjs/components/prism-cpp.js"
+import "prismjs/components/prism-csharp.js"
+import "prismjs/components/prism-diff.js"
+import "prismjs/components/prism-docker.js"
+import "prismjs/components/prism-git.js"
 import "prismjs/components/prism-go.js"
+import "prismjs/components/prism-graphql.js"
+import "prismjs/components/prism-handlebars.js"
+import "prismjs/components/prism-java.js"
+import "prismjs/components/prism-js-templates.js"
+import "prismjs/components/prism-kotlin.js"
+import "prismjs/components/prism-less.js"
+import "prismjs/components/prism-makefile.js"
+import "prismjs/components/prism-markdown.js"
+import "prismjs/components/prism-markup-templating.js"
+import "prismjs/components/prism-markup.js"
+import "prismjs/components/prism-objectivec.js"
+import "prismjs/components/prism-ocaml.js"
+import "prismjs/components/prism-python.js"
+import "prismjs/components/prism-reason.js"
+import "prismjs/components/prism-rust.js"
+import "prismjs/components/prism-sass.js"
+import "prismjs/components/prism-scss.js"
+import "prismjs/components/prism-solidity.js"
+import "prismjs/components/prism-sql.js"
+import "prismjs/components/prism-stylus.js"
+import "prismjs/components/prism-swift.js"
+import "prismjs/components/prism-wasm.js"
+import "prismjs/components/prism-yaml.js"
+import { ReactNode, useEffect } from "react"
+import useScheme from "src/hooks/useScheme"
+import Scripts from "src/layouts/RootLayout/Scripts"
+import Header from "./Header"
+import { ThemeProvider } from "./ThemeProvider"
+import useGtagEffect from "./useGtagEffect"
 
 type Props = {
   children: ReactNode
@@ -48,15 +47,15 @@ const RootLayout = ({ children }: Props) => {
   const [scheme] = useScheme()
   useGtagEffect()
   useEffect(() => {
-    Prism.highlightAll();
-  }, []);
+    Prism.highlightAll()
+  }, [])
 
   return (
     <ThemeProvider scheme={scheme}>
       <Scripts />
       {/* // TODO: replace react query */}
       {/* {metaConfig.type !== "Paper" && <Header />} */}
-      <Header fullWidth={false} />
+      <Header />
       <StyledMain>{children}</StyledMain>
     </ThemeProvider>
   )
