+++ src/routes/Detail/PostDetail/CommentBox/Utterances.tsx
@@ -1,8 +1,7 @@
-import { CONFIG } from "site.config"
-import { useEffect } from "react"
-import styled from "@emotion/styled"
-import useScheme from "src/hooks/useScheme"
 import { useRouter } from "next/router"
+import { useEffect } from "react"
+import { CONFIG } from "site.config"
+import useScheme from "src/hooks/useScheme"
 
 //TODO: useRef?
 
@@ -36,17 +35,11 @@ const Utterances: React.FC<Props> = ({ issueTerm }) => {
   }, [scheme, router])
   return (
     <>
-      <StyledWrapper id="comments">
+      <div id="comments">
         <div className="utterances-frame"></div>
-      </StyledWrapper>
+      </div>
     </>
   )
 }
 
 export default Utterances
-
-const StyledWrapper = styled.div`
-  @media (min-width: 768px) {
-    margin-left: -4rem;
-  }
-`
