+++ src/routes/Feed/Content/PostList/index.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router"
 import React, { useEffect, useState } from "react"
-import PostCard from "src/routes/Feed/PostList/PostCard"
 import { DEFAULT_CATEGORY } from "src/constants"
 import usePostsQuery from "src/hooks/usePostsQuery"
+import PostCard from "../../components/PostCard"
 
 type Props = {
   q: string
