+++ src/routes/Feed/Content/styles.ts
@@ -1,16 +1,25 @@
 import { css } from "@emotion/react"
+import { respondMobile } from "src/styles"
 
 export const wrapperStyle = css`
-  grid-column: span 12 / span 12;
+  grid-column: span 7 / span 7;
 
-  @media (min-width: 1024px) {
-    grid-column: span 7 / span 7;
+  ${respondMobile} {
+    grid-column: span 12 / span 12;
+  }
+`
+
+export const tagListStyle = css`
+  display: none;
+
+  ${respondMobile} {
+    display: block;
   }
 `
 
 export const footerStyle = css`
-  padding-bottom: 2rem;
-  @media (min-width: 1024px) {
-    display: none;
+  display: none;
+  ${respondMobile} {
+    padding-bottom: 2rem;
   }
 `
