+++ src/routes/Feed/LeftSidebar/styles.ts
@@ -1,21 +1,21 @@
 import { css } from "@emotion/react"
-import { HEADER_HEIGHT } from "../styles"
+import { respondMobile, variables } from "src/styles"
 
 export const wrapperStyle = css`
-  height: calc(100vh - ${HEADER_HEIGHT}px);
-  display: none;
+  height: calc(100vh - ${variables.headerHeight + 16}px);
   overflow: scroll;
   position: sticky;
   grid-column: span 2 / span 2;
-  top: ${HEADER_HEIGHT - 10}px;
+  top: ${variables.headerHeight + 16 - 10}px;
 
   scrollbar-width: none;
   -ms-overflow-style: none;
   &::-webkit-scrollbar {
     display: none;
   }
+  display: block;
 
-  @media (min-width: 1024px) {
-    display: block;
+  ${respondMobile} {
+    display: none;
   }
 `
